Defer authenticated shell work
