+++ apps/web/app/layout.tsx
@@ -2,10 +2,6 @@ import type { Metadata, Viewport } from "next";
 import { Inter, JetBrains_Mono } from "next/font/google";
 import { Agentation } from "agentation";
 import Script from "next/script";
-import { ConsentAwareAnalytics } from "@/components/providers/ConsentAwareAnalytics";
-import { PostHogClientProvider } from "@/components/providers/PostHogProvider";
-import { QueryProvider } from "@/components/providers/QueryProvider";
-import { ThemeProvider } from "@/components/providers/ThemeProvider";
 import { getThemeBootstrapScript } from "@/lib/theme";
 import "./globals.css";
 
@@ -126,12 +122,7 @@ export default function RootLayout({
         >
           Skip to content
         </a>
-        <PostHogClientProvider>
-          <QueryProvider>
-            <ThemeProvider>{children}</ThemeProvider>
-          </QueryProvider>
-        </PostHogClientProvider>
-        <ConsentAwareAnalytics />
+        {children}
         {process.env.NODE_ENV === "development" && <Agentation />}
       </body>
     </html>
